extract system info class to its own file
